wip: Implementing spot light shadow
