+++ EngineSIU/EngineSIU/editor.ini
@@ -1,5 +1,5 @@
 SplitterH.SplitRatio=0.500000
-PerspectiveCameraLocY0=-8.846296
+PerspectiveCameraLocY0=-0.657701
 PerspectiveCameraLocX3=0.000000
 PerspectiveCameraRotY1=0.000000
 SplitterV.SplitRatio=0.500000
@@ -8,25 +8,25 @@ ShowFlag1=31
 CameraSpeedSetting0=1
 CameraSpeedScalar0=0.100000
 PerspectiveCameraRotY2=0.000000
-PerspectiveCameraLocX0=13.070301
+PerspectiveCameraLocX0=-8.954199
 GridSize0=10.000000
-PerspectiveCameraLocZ0=21.239395
+PerspectiveCameraLocZ0=10.780217
 ShowFlag3=31
 CameraSpeedSetting2=1
 PerspectiveCameraRotX0=0.000000
 PerspectiveCameraLocX2=0.000000
-PerspectiveCameraRotY0=50.100815
-PerspectiveCameraRotZ0=-551.587585
+PerspectiveCameraRotY0=66.200829
+PerspectiveCameraRotZ0=-1087.757080
 PerspectiveCameraRotX3=0.000000
 CameraSpeedSetting1=1
 ShowFlag0=31
-ViewMode0=0
+ViewMode0=2
 ViewportType0=0
 CameraSpeedScalar1=1.000000
 GridSize1=10.000000
 PerspectiveCameraRotY3=0.000000
 PerspectiveCameraLocX1=0.000000
-ScreenHeight=961
+ScreenHeight=1137
 PerspectiveCameraLocY1=0.000000
 PerspectiveCameraLocZ1=0.000000
 CameraSpeedSetting3=1
@@ -40,7 +40,7 @@ GridSize2=10.000000
 PerspectiveCameraLocY2=0.000000
 PerspectiveCameraLocZ2=0.000000
 PerspectiveCameraRotZ2=0.000000
-ScreenWidth=1284
+ScreenWidth=1920
 ViewMode2=0
 ViewportType2=3
 CameraSpeedScalar3=1.000000
