+++ EngineSIU/EngineSIU/imgui.ini
@@ -3,12 +3,12 @@ Pos=60,60
 Size=400,400
 
 [Window][Outliner]
-Pos=1032,5
-Size=250,288
+Pos=1541,5
+Size=378,341
 
 [Window][Detail]
-Pos=1032,303
-Size=250,624
+Pos=1541,356
+Size=378,739
 
 [Window][Console]
 Pos=0,-30
@@ -17,9 +17,13 @@ Collapsed=1
 
 [Window][Control Panel]
 Pos=1,1
-Size=1027,50
+Size=1536,50
 
 [Window][Stat Overlay]
 Pos=0,0
 Size=32,32
+
+[Window][Menu]
+Pos=10,55
+Size=135,170
